# api-contract/react-client/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Baraka platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  role: string;
  accountType: string;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
}

export type SignupInputAccountType = typeof SignupInputAccountType[keyof typeof SignupInputAccountType];


export const SignupInputAccountType = {
  donor: 'donor',
  charity: 'charity',
} as const;

export interface SignupInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @maxLength 255 */
  email: string;
  /**
     * @minLength 10
     * @maxLength 200
     */
  password: string;
  /** @maxLength 32 */
  phone?: string;
  accountType?: SignupInputAccountType;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface Organization {
  id: number;
  name: string;
  /** @nullable */
  nameEn?: string | null;
  category: string;
  governorate: string;
  description: string;
  /** @nullable */
  descriptionEn?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  verified: boolean;
  createdAt: string;
}

export interface OrganizationInput {
  /** @minLength 1 */
  name: string;
  nameEn?: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  governorate: string;
  /** @minLength 1 */
  description: string;
  descriptionEn?: string;
  imageUrl?: string;
  phone?: string;
  email?: string;
  address?: string;
  lat?: number;
  lng?: number;
  verified?: boolean;
}

export interface OrganizationUpdate {
  /** @minLength 1 */
  name?: string;
  nameEn?: string;
  category?: string;
  governorate?: string;
  description?: string;
  descriptionEn?: string;
  imageUrl?: string;
  phone?: string;
  email?: string;
  address?: string;
  lat?: number;
  lng?: number;
  verified?: boolean;
}

export interface Beneficiary {
  id: number;
  name: string;
  story: string;
  /** @nullable */
  organizationId?: number | null;
  needAmount: number;
  raisedAmount: number;
  urgent: boolean;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface BeneficiaryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  story: string;
  organizationId?: number;
  /** @minimum 0 */
  needAmount?: number;
  urgent?: boolean;
  imageUrl?: string;
}

export interface BeneficiaryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  story?: string;
  organizationId?: number;
  /** @minimum 0 */
  needAmount?: number;
  urgent?: boolean;
  imageUrl?: string;
}

export interface Donation {
  id: number;
  donorId: number;
  /** @nullable */
  donorName?: string | null;
  /** @nullable */
  organizationId?: number | null;
  /** @nullable */
  organizationName?: string | null;
  /** @nullable */
  beneficiaryId?: number | null;
  amount: number;
  /** @nullable */
  message?: string | null;
  donationType: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  walletProvider?: string | null;
  /** @nullable */
  walletPhone?: string | null;
  /** @nullable */
  cardLast4?: string | null;
  /** @nullable */
  cardName?: string | null;
  /** @nullable */
  itemDetails?: string | null;
  createdAt: string;
}

export type DonationInputDonationType = typeof DonationInputDonationType[keyof typeof DonationInputDonationType];


export const DonationInputDonationType = {
  money: 'money',
  food: 'food',
  clothes: 'clothes',
  other: 'other',
} as const;

export type DonationInputPaymentMethod = typeof DonationInputPaymentMethod[keyof typeof DonationInputPaymentMethod];


export const DonationInputPaymentMethod = {
  card: 'card',
  wallet: 'wallet',
  cash: 'cash',
} as const;

export type DonationInputWalletProvider = typeof DonationInputWalletProvider[keyof typeof DonationInputWalletProvider];


export const DonationInputWalletProvider = {
  zain: 'zain',
  orange: 'orange',
  umniah: 'umniah',
} as const;

export interface DonationInput {
  /** @minimum 0 */
  amount: number;
  organizationId?: number;
  beneficiaryId?: number;
  message?: string;
  donationType: DonationInputDonationType;
  paymentMethod?: DonationInputPaymentMethod;
  walletProvider?: DonationInputWalletProvider;
  walletPhone?: string;
  /**
     * @minLength 4
     * @maxLength 4
     */
  cardLast4?: string;
  cardName?: string;
  itemDetails?: string;
}

export interface Announcement {
  id: number;
  title: string;
  body: string;
  kind: string;
  /** @nullable */
  organizationId?: number | null;
  /** @nullable */
  organizationName?: string | null;
  createdAt: string;
}

export interface AnnouncementInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  body: string;
  kind?: string;
  organizationId?: number;
}

export interface Notification {
  id: number;
  userId: number;
  title: string;
  body: string;
  read: boolean;
  createdAt: string;
}

export interface HelpRequest {
  id: number;
  name: string;
  phone: string;
  governorate: string;
  aidType: string;
  description: string;
  status: string;
  /** @nullable */
  adminNote?: string | null;
  /** @nullable */
  reviewedBy?: number | null;
  /** @nullable */
  reviewedAt?: string | null;
  createdAt: string;
}

export interface HelpRequestInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 1 */
  governorate: string;
  /** @minLength 1 */
  aidType: string;
  /** @minLength 5 */
  description: string;
}

export interface RegistrationRequest {
  id: number;
  orgName: string;
  governorate: string;
  category: string;
  /** @nullable */
  regNumber?: string | null;
  about: string;
  contactName: string;
  contactPhone: string;
  /** @nullable */
  email?: string | null;
  status: string;
  /** @nullable */
  adminNote?: string | null;
  /** @nullable */
  reviewedBy?: number | null;
  /** @nullable */
  reviewedAt?: string | null;
  createdAt: string;
}

export interface RegistrationRequestInput {
  /** @minLength 2 */
  orgName: string;
  /** @minLength 1 */
  governorate: string;
  /** @minLength 1 */
  category: string;
  regNumber?: string;
  /** @minLength 5 */
  about: string;
  /** @minLength 2 */
  contactName: string;
  /** @minLength 7 */
  contactPhone: string;
  email?: string;
}

export type RequestReviewStatus = typeof RequestReviewStatus[keyof typeof RequestReviewStatus];


export const RequestReviewStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface RequestReview {
  status: RequestReviewStatus;
  adminNote?: string;
}

export interface PlatformStats {
  totalRaised: number;
  donationCount: number;
  donorCount: number;
  organizationCount: number;
  beneficiaryCount: number;
}