# api-contract/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Baraka platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      operationId: deleteOrganization
      tags: [organizations]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /beneficiaries:
    get:
      operationId: listBeneficiaries
      tags: [beneficiaries]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Beneficiary"
    post:
      operationId: createBeneficiary
      tags: [beneficiaries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeneficiaryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Beneficiary"
  /beneficiaries/{id}:
    patch:
      operationId: updateBeneficiary
      tags: [beneficiaries]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeneficiaryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Beneficiary"
    delete:
      operationId: deleteBeneficiary
      tags: [beneficiaries]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /help-requests:
    get:
      operationId: listHelpRequests
      tags: [help-requests]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HelpRequest"
    post:
      operationId: createHelpRequest
      tags: [help-requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelpRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
  /help-requests/{id}:
    patch:
      operationId: updateHelpRequest
      tags: [help-requests]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestReview"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
...
components:
  schemas:
...
    AuthUser:
      type: object
      required: [id, name, email, role, accountType, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        accountType: { type: string }
        phone: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    SignupInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        email: { type: string, format: email, maxLength: 255 }
        # Minimum 10 chars (NIST 800-63B recommends >=8; we use 10 for charity-data context).
        # No complexity requirement — length beats complexity per OWASP guidance.
        password: { type: string, minLength: 10, maxLength: 200 }
        phone: { type: string, maxLength: 32 }
        accountType: { type: string, enum: [donor, charity] }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string, minLength: 1 }
    Organization:
      type: object
      required: [id, name, category, governorate, description, verified, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        nameEn: { type: ["string", "null"] }
        category: { type: string }
        governorate: { type: string }
        description: { type: string }
        descriptionEn: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        verified: { type: boolean }
        createdAt: { type: string, format: date-time }
    OrganizationInput:
      type: object
      required: [name, category, governorate, description]
      properties:
        name: { type: string, minLength: 1 }
        nameEn: { type: string }
        category: { type: string, minLength: 1 }
        governorate: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        descriptionEn: { type: string }
        imageUrl: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        lat: { type: number }
        lng: { type: number }
        verified: { type: boolean }
    OrganizationUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        nameEn: { type: string }
        category: { type: string }
        governorate: { type: string }
        description: { type: string }
        descriptionEn: { type: string }
        imageUrl: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        lat: { type: number }
        lng: { type: number }
        verified: { type: boolean }
    Beneficiary:
      type: object
      required: [id, name, story, needAmount, raisedAmount, urgent, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        story: { type: string }
        organizationId: { type: ["integer", "null"] }
        needAmount: { type: integer }
        raisedAmount: { type: integer }
        urgent: { type: boolean }
        imageUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    BeneficiaryInput:
      type: object
      required: [name, story]
      properties:
        name: { type: string, minLength: 1 }
        story: { type: string, minLength: 1 }
        organizationId: { type: integer }
        needAmount: { type: integer, minimum: 0 }
        urgent: { type: boolean }
        imageUrl: { type: string }
    BeneficiaryUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        story: { type: string, minLength: 1 }
        organizationId: { type: integer }
        needAmount: { type: integer, minimum: 0 }
        urgent: { type: boolean }
        imageUrl: { type: string }
...
    Announcement:
      type: object
      required: [id, title, body, kind, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        body: { type: string }
        kind: { type: string }
        organizationId: { type: ["integer", "null"] }
        organizationName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    AnnouncementInput:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 1 }
        body: { type: string, minLength: 1 }
        kind: { type: string }
        organizationId: { type: integer }
    Notification:
      type: object
      required: [id, userId, title, body, read, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        title: { type: string }
        body: { type: string }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
    HelpRequest:
      type: object
      required: [id, name, phone, governorate, aidType, description, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        governorate: { type: string }
        aidType: { type: string }
        description: { type: string }
        status: { type: string }
        adminNote: { type: ["string", "null"] }
        reviewedBy: { type: ["integer", "null"] }
        reviewedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    HelpRequestInput:
      type: object
      required: [name, phone, governorate, aidType, description]
      properties:
        name: { type: string, minLength: 2 }
        phone: { type: string, minLength: 7 }
        governorate: { type: string, minLength: 1 }
        aidType: { type: string, minLength: 1 }
        description: { type: string, minLength: 5 }
...
    RequestReview:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, approved, rejected] }
        adminNote: { type: string }
...