# api-contract/zod/src/generated/types/organization.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Baraka platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Organization {
  id: number;
  name: string;
  /** @nullable */
  nameEn?: string | null;
  category: string;
  governorate: string;
  description: string;
  /** @nullable */
  descriptionEn?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  verified: boolean;
  createdAt: Date;
}