# api-contract/zod/src/generated/types/organizationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Baraka platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OrganizationInput {
  /** @minLength 1 */
  name: string;
  nameEn?: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  governorate: string;
  /** @minLength 1 */
  description: string;
  descriptionEn?: string;
  imageUrl?: string;
  phone?: string;
  email?: string;
  address?: string;
  lat?: number;
  lng?: number;
  verified?: boolean;
}